0.2 - completely untested features!  Living dangerously!
Added Doors that open and close when they receive a signal.  These need to be tied to the global flags so they can stay open.
Added the Lightning Upgrade "enemy" to the EnemyFactory.  It should now actually appear on the maps.  If I had a map that used it.  Which I don't.