Bunches of changes.  Time crunch, so I don't even remember what I did.
Added Cutscene state.
Terminals can set flags.
Added the squasher optional boss.
Added the Boost plus and made it actually work.
Airblade is ready to be put in the game.
Playtesting and more notes taken.
Fixed nasty bug that makes the game unplayable when a saved game doesn't exist (which is the first time...)